lua: Renamed lua.{h,c} to lua-dest.{h,c}
Also did some cleanup around included headers.
